Couple updates for persisentTabs
Made it so if an error is thrown when trying to change the tab, we still
execute the code to remember the choice.

Made it so you can override the cookie value we set.  This is used on
the videos/create page since we want to sometimes navigate users to the
login page, but store the create public video choice in the persisentTab
cookie.

Removed some logging statements.